+++ tests/test_snapshot.py
@@ -4,7 +4,7 @@ from freezegun import freeze_time
 from lxml import etree
 from pathlib import Path
 
-from eicr_anonymization.anonimizer import Anonymizer
+from eicr_anonymization.anonymizer import Anonymizer
 from eicr_anonymization.anonymize_eicr import anonymize_eicr_file, xml_tree_to_str
 
 @freeze_time('2025-01-10 09:30:30')
